Consolidate switch port tests and resolve module collisions
    - Consolidate `tests/switch/test_switch_port.py` and `tests/switch/test_switch_port_entity.py` into `tests/switch/test_meraki_switch_port_switch.py` to prevent redundancy and test collection errors.
    - Rename `tests/sensor/device/test_switch_port.py` to `tests/sensor/device/test_meraki_switch_port_sensor.py` to resolve module naming collision.
    - Rename `tests/binary_sensor/test_switch_port.py` to `tests/binary_sensor/test_meraki_switch_port_binary_sensor.py` to resolve module naming collision.
    - Fix linting issue in `tests/switch/test_meraki_switch_port_switch.py`.
