+++ tests/binary_sensor/test_meraki_switch_port_binary_sensor.py
