+++ tests/core/test_entity_naming.py
@@ -1,9 +1,14 @@
 """Tests for entity naming."""
 
 from unittest.mock import MagicMock
+
 import pytest
-from custom_components.meraki_ha.core.entities.meraki_vlan_entity import MerakiVLANEntity
+
+from custom_components.meraki_ha.core.entities.meraki_vlan_entity import (
+    MerakiVLANEntity,
+)
 from custom_components.meraki_ha.types import MerakiNetwork
+
 
 @pytest.fixture
 def mock_coordinator():
