+++ tests/test_device_trigger.py
@@ -1,15 +1,7 @@
 """The tests for Meraki device triggers."""
 from __future__ import annotations
 
-from unittest.mock import AsyncMock
-
 import pytest
-from pytest_homeassistant_custom_component.common import (
-    MockConfigEntry,
-    async_get_device_automations,
-    async_mock_service,
-)
-
 from homeassistant.components import automation
 from homeassistant.components.device_automation import DeviceAutomationType
 from homeassistant.const import (
@@ -21,6 +13,11 @@ from homeassistant.const import (
 from homeassistant.core import HomeAssistant, ServiceCall
 from homeassistant.helpers import device_registry as dr
 from homeassistant.setup import async_setup_component
+from pytest_homeassistant_custom_component.common import (
+    MockConfigEntry,
+    async_get_device_automations,
+    async_mock_service,
+)
 
 from custom_components.meraki_ha.const import DOMAIN, EVENT_MERAKI_WEBHOOK_ALERT
